# contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Jitong Server Contract
  version: 1.0.0
  description: Baseline operational contract. Business endpoints are added in later tracer bullets.
paths:
...
  /api/v1/admin/audit-logs:
    get:
      operationId: listAuditLogs
      summary: Query content-free security audit metadata
      security: [{ adminApiKey: [] }]
      parameters:
        - name: eventType
          in: query
          schema: { type: string }
        - name: outcome
          in: query
          schema: { type: string, enum: [SUCCEEDED, REJECTED, FAILED] }
        - name: actorUserId
          in: query
          schema: { type: string, format: uuid }
        - name: subjectType
          in: query
          schema: { type: string }
        - name: subjectId
          in: query
          schema: { type: string, format: uuid }
        - name: requestId
          in: query
          schema: { type: string, format: uuid }
        - name: occurredBefore
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        '200':
          description: Content-free audit metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditLogRecord' }
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
...
  /api/v1/conversations/{conversationId}/ai/consent:
    get:
      operationId: getC2cAiConsent
      summary: Read the authenticated user's C2C private AI consent
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ConversationId' }
      responses:
        '200':
          description: Current consent and effective bilateral state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiConsentResponse' }
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
    patch:
      operationId: updateC2cAiConsent
      summary: Enable or disable the authenticated user's C2C private AI consent
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ConversationId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          description: Updated consent and effective bilateral state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiConsentResponse' }
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
  /api/v1/conversations/{conversationId}/ai/summary:
    post:
      operationId: enqueueC2cAiSummary
      summary: Queue a private C2C summary for the authenticated user
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ConversationId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [requestId]
              properties:
                requestId: { type: string, format: uuid }
                afterSeq: { type: [integer, 'null'], format: int64, minimum: 0 }
                untilSeq: { type: [integer, 'null'], format: int64, minimum: 0 }
      responses:
        '200':
          description: Idempotent queued private summary job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiJobResponse' }
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
        '429':
          $ref: '#/components/responses/ApiError'
  /api/v1/conversations/{conversationId}/ai/smart-replies:
    post:
      operationId: enqueueC2cAiSmartReplies
      summary: Queue three private editable C2C reply drafts for the authenticated user
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ConversationId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [requestId]
              properties:
                requestId: { type: string, format: uuid }
      responses:
        '200':
          description: Idempotent private smart-reply job; drafts are never sent automatically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiJobResponse' }
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
        '429':
          $ref: '#/components/responses/ApiError'
  /api/v1/conversations/{conversationId}/ai/extract:
    post:
      operationId: enqueueC2cAiExtraction
      summary: Queue private action-item and key-fact extraction from selected C2C messages
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ConversationId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [requestId, messageIds]
              properties:
                requestId: { type: string, format: uuid }
                messageIds:
                  type: array
                  minItems: 1
                  maxItems: 200
                  uniqueItems: true
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: Idempotent private information-extraction job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiJobResponse' }
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
        '429':
          $ref: '#/components/responses/ApiError'
  /api/v1/ai/jobs:
    get:
      operationId: listPrivateAiJobs
      summary: List unexpired private AI job statuses owned by the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Private AI job statuses used for authoritative device restore
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiJobStatusResponse' }
        '401':
          $ref: '#/components/responses/ApiError'
  /api/v1/ai/jobs/{jobId}:
    get:
      operationId: getPrivateAiJob
      summary: Read one private AI job owned by the authenticated user
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/JobId' }
      responses:
        '200':
          description: Private AI job and optional structured result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiJobResponse' }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deletePrivateAiJob
      summary: Delete the authenticated user's private AI job result
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/JobId' }
      responses:
        '204': { description: Private AI job result deleted }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /api/v1/ai/artifacts:
    get:
      operationId: listPrivateAiArtifacts
      summary: List unexpired private AI artifacts owned by the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Private AI artifacts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiArtifactResponse' }
        '401':
          $ref: '#/components/responses/ApiError'
  /api/v1/ai/artifacts/{artifactId}:
    delete:
      operationId: deletePrivateAiArtifact
      summary: Delete one private AI artifact owned by the authenticated user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: artifactId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Private AI artifact deleted }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
  /api/v1/ai/action-items:
    get:
      operationId: listPrivateAiActionItems
      summary: List private action items owned by the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Private action items retained until completion or deletion
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiActionItemResponse' }
        '401':
          $ref: '#/components/responses/ApiError'
  /api/v1/ai/action-items/{actionItemId}:
    patch:
      operationId: updatePrivateAiActionItem
      summary: Mark an owned private action item open or completed
      security: [{ bearerAuth: [] }]
      parameters:
        - name: actionItemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status: { type: string, enum: [OPEN, COMPLETED] }
      responses:
        '200':
          description: Updated private action item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiActionItemResponse' }
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
    delete:
      operationId: deletePrivateAiActionItem
      summary: Hard-delete an owned private action item
      security: [{ bearerAuth: [] }]
      parameters:
        - name: actionItemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Private action item deleted }
        '401':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  parameters:
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    AuditLogRecord:
      type: object
      additionalProperties: false
      required:
        [id, eventType, outcome, actorUserId, actorDeviceId, subjectType,
         subjectId, requestId, errorCode, occurredAt]
      properties:
        id: { type: string, format: uuid }
        eventType: { type: string }
        outcome: { type: string, enum: [SUCCEEDED, REJECTED, FAILED] }
        actorUserId: { type: [string, 'null'], format: uuid }
        actorDeviceId: { type: [string, 'null'], format: uuid }
        subjectType: { type: [string, 'null'] }
        subjectId: { type: [string, 'null'], format: uuid }
        requestId: { type: [string, 'null'], format: uuid }
        errorCode: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }
...
    AiConsentResponse:
      type: object
      additionalProperties: false
      required: [version, conversationId, userId, enabled, enabledForBoth, policyVersion]
      properties:
        version: { const: 1 }
        conversationId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        enabled: { type: boolean }
        enabledForBoth: { type: boolean }
        policyVersion: { type: integer, format: int64, minimum: 1 }
    AiSummary:
      type: object
      additionalProperties: false
      required: [overview, keyPoints, decisions, openQuestions, sourceMessageIds]
      properties:
        overview: { type: string, minLength: 1, maxLength: 4000 }
        keyPoints:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1 }
        decisions:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1 }
        openQuestions:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1 }
        sourceMessageIds:
          type: array
          items: { type: string, format: uuid }
    AiSmartReplies:
      type: object
      additionalProperties: false
      required: [replies]
      properties:
        replies:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: object
            additionalProperties: false
            required: [text, tone]
            properties:
              text: { type: string, minLength: 1, maxLength: 2000 }
              tone: { type: string, minLength: 1, maxLength: 40 }
    AiExtraction:
      type: object
      additionalProperties: false
      required: [actionItems, keyFacts]
      properties:
        actionItems:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [title, details, assigneeUserId, dueAt, priority, confidence, sourceMessageIds]
            properties:
              title: { type: string, minLength: 1, maxLength: 500 }
              details: { type: string, maxLength: 4000 }
              assigneeUserId: { type: [string, 'null'], format: uuid }
              dueAt: { type: [string, 'null'], format: date-time }
              priority: { type: string, enum: [LOW, MEDIUM, HIGH] }
              confidence: { type: number, minimum: 0, maximum: 1 }
              sourceMessageIds:
                type: array
                minItems: 1
                items: { type: string, format: uuid }
        keyFacts:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [category, content, confidence, sourceMessageIds]
            properties:
              category: { type: string, minLength: 1, maxLength: 80 }
              content: { type: string, minLength: 1, maxLength: 4000 }
              confidence: { type: number, minimum: 0, maximum: 1 }
              sourceMessageIds:
                type: array
                minItems: 1
                items: { type: string, format: uuid }
    AiJobStatusResponse:
      type: object
      additionalProperties: false
      required: [version, jobId, conversationId, kind, status, errorCode, createdAt, expiresAt]
      properties:
        version: { const: 1 }
        jobId: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        kind: { type: string, enum: [SUMMARY, SMART_REPLY, EXTRACTION] }
        status: { type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED] }
        errorCode: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    AiJobResponse:
      type: object
      additionalProperties: false
      required:
        [version, jobId, ownerUserId, requestingDeviceId, conversationId, requestId,
         kind, status, fromSeq, toSeq, model, promptVersion, errorCode,
         createdAt, startedAt, finishedAt, expiresAt, result]
      properties:
        version: { const: 1 }
        jobId: { type: string, format: uuid }
        ownerUserId: { type: string, format: uuid }
        requestingDeviceId: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        requestId: { type: string, format: uuid }
        kind: { type: string, enum: [SUMMARY, SMART_REPLY, EXTRACTION] }
        status: { type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED, EXPIRED] }
        fromSeq: { type: integer, format: int64, minimum: 1 }
        toSeq: { type: integer, format: int64, minimum: 1 }
        model: { type: string, minLength: 1 }
        promptVersion: { type: string, minLength: 1 }
        errorCode: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, 'null'], format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        expiresAt: { type: [string, 'null'], format: date-time }
        result:
          oneOf:
            - { $ref: '#/components/schemas/AiSummary' }
            - { $ref: '#/components/schemas/AiSmartReplies' }
            - { $ref: '#/components/schemas/AiExtraction' }
            - { type: 'null' }
    AiArtifactResponse:
      type: object
      additionalProperties: false
      required: [version, artifactId, jobId, conversationId, artifactType, content, createdAt, expiresAt]
      properties:
        version: { const: 1 }
        artifactId: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        artifactType: { type: string, enum: [SUMMARY, SMART_REPLY, EXTRACTION] }
        content:
          oneOf:
            - { $ref: '#/components/schemas/AiSummary' }
            - { $ref: '#/components/schemas/AiSmartReplies' }
            - { $ref: '#/components/schemas/AiExtraction' }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    AiActionItemResponse:
      type: object
      additionalProperties: false
      required:
        [version, actionItemId, sourceJobId, ownerUserId, conversationId,
         assigneeUserId, title, details, dueAt, priority, confidence,
         sourceMessageIds, status, createdAt, completedAt]
      properties:
        version: { const: 1 }
        actionItemId: { type: string, format: uuid }
        sourceJobId: { type: [string, 'null'], format: uuid }
        ownerUserId: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        assigneeUserId: { type: [string, 'null'], format: uuid }
        title: { type: string, minLength: 1, maxLength: 500 }
        details: { type: string, maxLength: 4000 }
        dueAt: { type: [string, 'null'], format: date-time }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        sourceMessageIds:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
        status: { type: string, enum: [OPEN, COMPLETED] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
  responses:
    ApiError:
      description: Versioned error response
      headers:
        X-Request-Id:
          description: UUIDv4 request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/json:
          schema:
            $ref: './schemas/error-response-v1.schema.json'